requirements.txt
alabaster==1.0.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
app-model==0.4.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
build==1.3.0
cachey==0.2.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
dask==2025.7.0
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
docstring_parser==0.17.0
docutils==0.21.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
flexcache==0.3
flexparser==0.4
fonttools==4.59.0
fqdn==1.5.1
freetype-py==2.5.1
fsspec==2025.7.0
h11==0.16.0
HeapDict==1.0.1
hsluv==5.0.4
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
imageio==2.37.0
imagesize==1.4.1
imgaug==0.4.0
in-n-out==0.2.1
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
lark==1.2.2
lazy_loader==0.4
lmdb==1.7.3
locket==1.0.0
magicgui==0.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mmocr==1.0.1
mpmath==1.3.0
napari==0.6.3
napari-console==0.1.3
napari-plugin-engine==0.2.0
napari-svg==0.2.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.5
notebook_shim==0.2.4
npe2==0.7.9
numpy==2.2.6
numpydoc==1.9.0
omegaconf==2.3.0
opencv-python==4.12.0.88
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pillow==11.3.0
Pint==0.24.4
platformdirs==4.3.8
pooch==1.8.2
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psutil==7.0.0
psygnal==0.14.0
pure_eval==0.2.3
pyclipper==1.3.0.post6
pycocotools==2.0.10
pyconify==0.2.1
pycparser==2.22
pydantic==2.11.7
pydantic-compat==0.1.2
pydantic_core==2.33.2
Pygments==2.19.2
PyOpenGL==3.1.9
pyparsing==3.2.3
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
pywin32==311
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==27.0.1
qtconsole==5.6.1
QtPy==2.4.3
RapidFuzz==3.13.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
roman-numerals-py==3.1.0
rpds-py==0.27.0
safetensors==0.6.1
scikit-image==0.25.2
scipy==1.16.1
Send2Trash==1.8.3
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
soupsieve==2.7
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
superqt==0.7.5
sympy==1.14.0
terminado==0.18.1
tifffile==2025.6.11
tinycss2==1.4.0
tokenizers==0.21.4
tomli_w==1.2.0
toolz==1.0.0
torch==2.8.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.0
typer==0.16.0
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
vispy==0.15.2
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.14
wrapt==1.17.2
